added functionality to activate tenants based on the object
git-svn-id: http://svn/php/pimcore_plugins/onlineshop/trunk/www@317 6df6f5fa-e861-40d1-bdbc-1b55999331ee

+++ lib/OnlineShop/Framework/IndexService/Tenant/Worker.php
@@ -128,8 +128,12 @@ class OnlineShop_Framework_IndexService_Tenant_Worker {
 
     }
 
-
     public function deleteFromIndex(OnlineShop_Framework_ProductInterfaces_IIndexable $object){
+        if(!$this->tenantConfig->isActive($object)) {
+            Logger::info("Tenant is not active.");
+            return;
+        }
+
         $this->db->delete($this->tenantConfig->getTablename(), "o_id = " . $object->getId());
         $this->db->delete($this->tenantConfig->getRelationTablename(), "src = " . $this->db->quote($object->getId()));
         if($this->tenantConfig->getTenantRelationTablename()) {
@@ -138,6 +142,11 @@ class OnlineShop_Framework_IndexService_Tenant_Worker {
     }
 
     public function updateIndex(OnlineShop_Framework_ProductInterfaces_IIndexable $object) {
+        if(!$this->tenantConfig->isActive($object)) {
+            Logger::info("Tenant is not active.");
+            return;
+        }
+
 
         if($object->getOSDoIndexProduct() && $this->tenantConfig->inIndex($object)) {
             $a = Pimcore::inAdmin();
